Make version a var and add arr_build.sh
By change Version from and const to a var we are able to use
ldflags to set the version dynamically.

In the arr_build.sh file for example, we can set it with:
-ldflags="-X go.mozilla.org/sops/v3/version.Version=$VERSION"

arr_build.sh will buld the sops binary in the repo root
and will print the instructions for releasing it to gcs.